requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
